Remove Geckos (WebRTC) from the networking stack. Closes #156

+++ desktop/main/helpers/startThoriumServer.ts
@@ -1,14 +1,16 @@
-import {app, BrowserWindow} from "electron";
+import {app} from "electron";
 import path from "path";
-import {fork} from "child_process";
-import {ipcMain} from "electron-better-ipc";
-let restartCount = 0;
-export default function startThoriumServer(browserWindow: BrowserWindow) {
-  function startServer() {
-    const child = fork(path.join(app.getAppPath(), "dist/index.js"), [], {
+import {fork, ChildProcess} from "child_process";
+
+let child: ChildProcess | null = null;
+export async function startThoriumServer() {
+  return new Promise<void>(function startServer(resolve, reject) {
+    if (child) return;
+    child = fork(path.join(app.getAppPath(), "dist/index.js"), [], {
       env: {
         FORK: "1",
         NODE_ENV: "production",
+        NODE_PATH: path.join(app.getAppPath(), "dist"),
         ...process.env,
       },
       // execArgv: [
@@ -19,57 +21,32 @@ export default function startThoriumServer(browserWindow: BrowserWindow) {
       // ],
       silent: true,
     });
-
+    child.on("message", function (msg) {
+      if (msg === "ready") {
+        resolve();
+      } else {
+        reject();
+      }
+    });
     child.stdout?.on("data", function (data) {
       const message: string = data.toString();
-      ipcMain.callRenderer(browserWindow, "server_info", message);
+      console.log(message);
     });
     child.stderr?.on("data", function (data) {
       const error = data.toString();
       if (error.includes("DeprecationWarning: Buffer()")) return;
-      ipcMain.callRenderer(browserWindow, "server_info", error);
-    });
-    child.on("close", function (code) {
-      if (browserWindow && restartCount < 10) {
-        try {
-          ipcMain.callRenderer(
-            browserWindow,
-            "server_info",
-            `Server process closed. Restarting...`
-          );
-          startServer();
-        } catch {
-          // Do nothing, we're probably shutting down.
-        }
-      } else {
-        if (browserWindow) {
-          ipcMain.callRenderer(
-            browserWindow,
-            "server_info",
-            `Server process closed. Too many restarts. Closing Thorium Server.`
-          );
-        }
-        app.quit();
-        return;
-      }
-      restartCount++;
+      console.error(error);
     });
     child.on("error", function (err) {
-      ipcMain.callRenderer(
-        browserWindow,
-        "server_info",
-        `Error in server process: ${err.message}`
-      );
       console.error(err);
     });
 
     app.on("before-quit", () => {
-      child.kill();
+      if (child) child.kill();
     });
-  }
-  startServer();
+  });
 }
-
-setInterval(() => {
-  restartCount = Math.max(0, restartCount - 1);
-}, 10 * 1000);
+export function stopThoriumServer() {
+  if (child) child.kill();
+  child = null;
+}
